user todos

+++ app/Http/Controllers/TodosController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Todo;
+use App\User;
 use Illuminate\Http\Request;
 
 class TodosController extends Controller
@@ -14,7 +15,7 @@ class TodosController extends Controller
      */
     public function index()
     {
-        return Todo::all();
+        return Todo::where('user_id', auth()->user()->id)->get();
     }
 
     /**
@@ -30,7 +31,11 @@ class TodosController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()-user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed
+        ]);
 
         return response($todo, 201);
     }
@@ -44,6 +49,10 @@ class TodosController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if ($todo->user_id != auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $data = $request->validate([
             'title' => 'required|string',
             'completed' => 'required|boolean',
@@ -60,7 +69,7 @@ class TodosController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
 
         return response('Updated', 200);
     }
@@ -74,6 +83,10 @@ class TodosController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if ($todo->user_id != auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $todo->delete();
 
         return response('Deleted todo item', 200);
@@ -81,12 +94,27 @@ class TodosController extends Controller
 
     public function destroyCompleted(Request $request)
     {
-        $data = $request->validate([
+
+        $todosToDelelte = $request->todos;
+
+        $userTodoIds = auth()->user()->todos->map(function ($todo) {
+            return $todo->id;
+        });
+
+        $valid = collect($todosToDelelte)->every(function ($value, $key) use ($userTodoIds) {
+            return $userTodoIds->contains($value);
+        });
+
+        if (!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
+
+        $request->validate([
             'todo' => 'required|array',
         ]);
 
         Todo::destroy($request->todo);
 
-        return response('Deleted', 200);
+        return response()->json('Deleted', 200);
     }
 }
